ref: fix 149 Qodana code quality issues
  Address all code quality issues identified by Qodana JVM 2024.3 analysis:

  Security (6):
  - Update protobuf-java 3.24.0 → 4.28.2 (CVE-2024-7254)
  - Update commons-lang3 3.12.0 → 3.17.0 (CVE-2025-48924)

  Potential Bugs (4):
  - Remove dead code with wrong type in OneofConflictDetector
  - Add try-with-resources in SchemaDiffCli
  - Remove redundant assignments in MergedFieldContract, SchemaDiffEngine

  Code Cleanup (33):
  - Remove unused imports across 17 files

  Simplifications (36):
  - Simplify trivial if statements to direct returns (10 files)
  - Remove redundant conditions and dead code branches
  - Remove redundant type casts in CodeGenerationHelper
  - Add explicit (int) cast for stream().count() in MessageDiff

  Field Issues (5):
  - Add final modifier to collection fields in GeneratorConfig
  - Remove unused field in VersionMerger

  Deprecation (4):
  - Replace deprecated API calls with modern alternatives
  - Use ProtoSyntax enum instead of protobufMajorVersion()

  Code Duplication (6 handlers refactored):
  - Extract shared addMissingFieldExtract() utility methods
  - Refactor SignedUnsignedHandler, RepeatedSingleHandler, EnumEnumHandler,
    IntEnumHandler, PrimitiveMessageHandler, StringBytesHandler

  Documentation:
  - Add Qodana badge to README.md
  - Create QODANA_FIXES.md with detailed fix plan and progress

  All tests pass. No functional changes.

+++ proto-wrapper-core/src/main/java/space/alnovis/protowrapper/diff/model/MessageDiff.java
@@ -168,7 +168,7 @@ public record MessageDiff(
             count++;
         }
 
-        count += fieldChanges.stream().filter(FieldChange::isBreaking).count();
+        count += (int) fieldChanges.stream().filter(FieldChange::isBreaking).count();
         count += nestedMessageChanges.stream().mapToInt(MessageDiff::countBreakingChanges).sum();
         count += nestedEnumChanges.stream()
             .filter(EnumDiff::hasBreakingChanges)
